# lib/api-zod/src/generated/types/addNoteBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddNoteBody {
  type: string;
  sourceUrl?: string | null;
  sourceText: string;
  response?: string | null;
}