# lib/api-zod/src/generated/types/opinionResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonaNote } from "./personaNote";

export interface OpinionResponse {
  opinion: string;
  note: PersonaNote;
}