# lib/api-zod/src/generated/types/personaDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { PersonaDetailData } from "./personaDetailData";
import type { PersonaNote } from "./personaNote";

export interface PersonaDetail {
  id: string;
  name: string;
  role: string;
  company: string;
  data: PersonaDetailData;
  notes: PersonaNote[];
  createdAt: Date;
}