# lib/api-zod/src/generated/types/personaListItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface PersonaListItem {
  id: string;
  name: string;
  role: string;
  company: string;
  createdAt: Date;
}